// src/lib/types.d.ts
export type BlogPost = {
  title: string;
  categories: string[];
  thumbnail: string;
  summary: string;
  publishedDate: string;
  id: string;
  country: string;
};

export type PostPage = {
  post: BlogPost;
  markdown: string;
};